Create Sonoff_Mini1_Timer_Toggle.yaml
New configuration to act like a 3 way switch (toggle mode), with an added delay. 

I use this as a bathroom exhaust fan. The regular wall switch will turn the fan on for a set amount of time, then turn off. There are also direct interface to immediately turn on or off to be used via the automation console based on other sensors, etc.